ai-daily-report-github/requirements.txt
# AI日报系统 - Python依赖

# RSS解析
feedparser>=6.0.10

# 网页爬取
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# 正文提取
readability-lxml>=0.8.1

# AI服务 - DeepSeek API
openai>=1.12.0

# 时间处理
python-dateutil>=2.8.2
pytz>=2024.1

# 环境变量
python-dotenv>=1.0.0
